+++ CHANGELOG.md
@@ -5,8 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project mostly adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 1.0.0 - 2019-02-17
+## [2.0.0-beta.1] - 2019-04-07
+
+### Added
+- Tailwind 1.0.0 compatibility
+- Added options to enable/disable the utilities
+
+### Changed
+- Most of the config options have been moved to the `theme` and `variants` objects in your Tailwind config (see `README` for more info)
+- Container class names now always include the word `container`; you can however customize the suffix (the keys in the `theme.fluidContainer` object, where `default` means no suffix) and the component prefix (which still defaults to `c-`)
+
+## [1.0.0] - 2019-02-17
 
 Initial release
 
-[Unreleased]: https://github.com/benface/tailwindcss-fluid-container/compare/v1.0.0...HEAD
+[Unreleased]: https://github.com/benface/tailwindcss-fluid-container/compare/v2.0.0-beta.1...HEAD
+[2.0.0-beta.1]: https://github.com/benface/tailwindcss-fluid-container/compare/v1.0.0...v2.0.0-beta.1
+[1.0.0]: https://github.com/benface/tailwindcss-fluid-container/releases/tag/v1.0.0